--- server/aList/aListModel.js
@@ -1,7 +1,0 @@
-
-// Database model for the aList
-
-// Will include the schema and any logic that
-// needs to be done on model creation and/or save
-
-module.exports = {};
