--- server/aList/aListRoutes.js
@@ -1,9 +1,0 @@
-
-// Express router for the aList api endpoint
-
-var aListController = require('./aListController.js');
-
-module.exports = function(app) {
-  app.route('/')
-    .get(aListController.top25);
-};
