+++ server/api/twitter.js
@@ -1,6 +1,6 @@
 
 var OAuth2 = require('OAuth').OAuth2;
-var keys = require('../keys.js');
+var keys = require('../../keys.js');
 var https = require('https');
 
 var oauth2 = new OAuth2(keys.twitter.key, keys.twitter.secret, 'https://api.twitter.com/', null, 'oauth2/token', null);
@@ -23,8 +23,8 @@ oauth2.getOAuthAccessToken('', {grant_type: 'client_credentials'}, function (e, 
     });
 
     result.on('end', function() {
-      var tweets = JSON.parse(buffer);
-      console.log(tweets);
+      var tweetData = JSON.parse(buffer);
+      console.log(tweetData[0].trends);
     });
   });
 });
