--- server/people/peopleModel.js
@@ -1,7 +1,0 @@
-
-// Database model for people
-
-// Will include the schema and any logic that
-// needs to be done on model creation and/or save
-
-module.exports = {};
