+++ server/top/topRoutes.js
@@ -1,0 +1,9 @@
+
+// Express router for the top api endpoint
+
+var topController = require('./topController.js');
+
+module.exports = function(app) {
+  app.route('/')
+    .get(topController.top25);
+};
